+++ CHANGELOG.md
@@ -13,10 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added on-demand _AudioClip_ module
 * Added Compute Shader Variants support
 * Fixed over-reporting of built shader variants count
+* Fixed export of filtered/selected non-diagnostic issues
 * Fixed Build Report object name
 * Fixed text alignment and wrapping issues
+* Fixed build report steps text wrapping
 * Improved text search to match custom properties
-* Fix export of filtered/selected non-diagnostic issues
 
 ## [0.8.4-preview] - 2022-09-27
 * Added Packages module to report installed packages and dependencies